requirements.txt
google-genai
python-dotenv
fastapi
uvicorn[standard]
python-multipart
